requirements.txt
pydantic>=1.0,<2.0
requests
sqlalchemy
snowflake-snowpark-python